+++ libdebug.cpp
@@ -25,12 +25,12 @@ int32 scriptlib::debug_message(lua_State *L) {
 int32 scriptlib::debug_add_card(lua_State *L) {
 	check_param_count(L, 6);
 	duel* pduel = interpreter::get_duel_info(L);
-	int32 code = lua_tointeger(L, 1);
-	int32 owner = lua_tointeger(L, 2);
-	int32 playerid = lua_tointeger(L, 3);
-	int32 location = lua_tointeger(L, 4);
-	int32 sequence = lua_tointeger(L, 5);
-	int32 position = lua_tointeger(L, 6);
+	int32 code = (int32)lua_tointeger(L, 1);
+	int32 owner = (int32)lua_tointeger(L, 2);
+	int32 playerid = (int32)lua_tointeger(L, 3);
+	int32 location = (int32)lua_tointeger(L, 4);
+	int32 sequence = (int32)lua_tointeger(L, 5);
+	int32 position = (int32)lua_tointeger(L, 6);
 	int32 proc = lua_toboolean(L, 7);
 	if(owner != 0 && owner != 1)
 		return 0;
@@ -67,10 +67,10 @@ int32 scriptlib::debug_add_card(lua_State *L) {
 int32 scriptlib::debug_set_player_info(lua_State *L) {
 	check_param_count(L, 4);
 	duel* pduel = interpreter::get_duel_info(L);
-	uint32 playerid = lua_tointeger(L, 1);
-	uint32 lp = lua_tointeger(L, 2);
-	uint32 startcount = lua_tointeger(L, 3);
-	uint32 drawcount = lua_tointeger(L, 4);
+	uint32 playerid = (uint32)lua_tointeger(L, 1);
+	uint32 lp = (uint32)lua_tointeger(L, 2);
+	uint32 startcount = (uint32)lua_tointeger(L, 3);
+	uint32 drawcount = (uint32)lua_tointeger(L, 4);
 	if(playerid != 0 && playerid != 1)
 		return 0;
 	pduel->game_field->player[playerid].lp = lp;
@@ -109,8 +109,8 @@ int32 scriptlib::debug_pre_add_counter(lua_State *L) {
 	check_param_count(L, 2);
 	check_param(L, PARAM_TYPE_CARD, 1);
 	card* pcard = *(card**) lua_touserdata(L, 1);
-	uint32 countertype = lua_tointeger(L, 2);
-	uint32 count = lua_tointeger(L, 3);
+	uint32 countertype = (uint32)lua_tointeger(L, 2);
+	uint32 count = (uint32)lua_tointeger(L, 3);
 	uint16 cttype = countertype;
 	if((countertype & COUNTER_NEED_ENABLE) && !(countertype & COUNTER_NEED_PERMIT))
 		cttype &= 0xfff;
@@ -129,7 +129,7 @@ int32 scriptlib::debug_pre_add_counter(lua_State *L) {
 int32 scriptlib::debug_reload_field_begin(lua_State *L) {
 	check_param_count(L, 1);
 	duel* pduel = interpreter::get_duel_info(L);
-	uint32 flag = lua_tointeger(L, 1);
+	uint32 flag = (uint32)lua_tointeger(L, 1);
 	pduel->clear();
 	pduel->game_field->core.duel_options = flag;
 	return 0;
@@ -149,7 +149,7 @@ int32 scriptlib::debug_set_ai_name(lua_State *L) {
 	duel* pduel = interpreter::get_duel_info(L);
 	pduel->write_buffer8(MSG_AI_NAME);
 	const char* pstr = lua_tostring(L, 1);
-	int len = strlen(pstr);
+	int32 len = (int32)strlen(pstr);
 	if(len > 100)
 		len = 100;
 	pduel->write_buffer16(len);
@@ -165,7 +165,7 @@ int32 scriptlib::debug_show_hint(lua_State *L) {
 	duel* pduel = interpreter::get_duel_info(L);
 	pduel->write_buffer8(MSG_SHOW_HINT);
 	const char* pstr = lua_tostring(L, 1);
-	int len = strlen(pstr);
+	int32 len = (int32)strlen(pstr);
 	if(len > 1024)
 		len = 1024;
 	pduel->write_buffer16(len);
